Revert "wt88047: camera: Partial update HAL post processing"
This reverts commit a71f84f5de4daaf7e81be821da0f401c32e685f6.

Conflicts:
	camera/QCamera2/stack/mm-jpeg-interface/Android.mk

+++ camera/QCamera2/HAL/QCameraMem.cpp
@@ -178,7 +178,7 @@ int QCameraMemory::getSize(int index) const
  *
  * RETURN     : number of buffers allocated
  *==========================================================================*/
-uint8_t QCameraMemory::getCnt() const
+int QCameraMemory::getCnt() const
 {
     return mBufferCount;
 }
